requirements.txt
gTTS
moviepy
opencv-python
numpy
requests
newspaper3k
GoogleNews
tqdm
torch
torchvision
pygame
lxml_html_clean
scipy
pillow
pyttsx3